% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fPQRHMM.R
\name{check_lossTE}
\alias{check_lossTE}
\title{Check loss}
\usage{
check_lossTE(beta_pred, X_test, Y_test, q)
}
\arguments{
\item{beta_pred}{coefficients for model}

\item{X_test}{test set for predictor}

\item{Y_test}{test set for response}

\item{q}{quantile you are predicting}
}
\value{
check_lossTE= check loss error of model given by beta_pred on test set given by X_test and Y_test
}
\description{
This function calculates check loss error for a fitted quantile regression model.
}
\examples{
 X_0=MASS::mvrnorm(100, mu=rep(0,100), Sigma=diag(100))
 beta=rep(1,100)
 Y_0= X_0\%*\%beta+rnorm(100)
 fPQR_mod= fPQR_fit(X_0, Y_0, 0.5)
 est_coef=fPQR_mod[[1]]
 check_losserror=check_lossTE(est_coef, X_0, Y_0,0.5)
}
