% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{fPQR_fit}
\alias{fPQR_fit}
\title{Quantile regression  using fPQR+HMM}
\usage{
fPQR_fit(X_0, Y_0, q)
}
\arguments{
\item{X_0}{predictor matrix}

\item{Y_0}{response matrix}

\item{q}{quantile you are trying to predict}
}
\value{
A list containing the following
\itemize{
  \item  rbind(intercept,beta): estimated regression coefficients for X_0
  \item  final_mod: Final fitted model for quantile regression of Y_0 on score matrix
  \item decomp_time: Amount of time for fPQR to decompose X
  \item select_time: Amount of time for QREM_vs (aka HMM) to run
  \item fit_time: Amount of time for QREM to run
  \item convert_time: Amount of time to convert coefficients for score matrix into coefficients for X_0
 }
}
\description{
This function is used to fit quantile regression models using the fPQR and HMM algorithms and is based
 off of code written by Civieta et al. in 2021.
}
\examples{
X_0= MASS::mvrnorm(100, mu=rep(0,100), Sigma=diag(100))
 beta=rep(1,100)
 Y_0= X_0\%*\%beta+rnorm(100)
 fPQR_mod= fPQR_fit(X_0, Y_0, 0.5)
}
