% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{qcov_li}
\alias{qcov_li}
\title{Quantile Covariance}
\usage{
qcov_li(X, Y, q)
}
\arguments{
\item{X}{covariate matrix}

\item{Y}{response matrix}

\item{q}{quantile you are predicting}
}
\value{
qcov_li: quantile covariance of X and Y
}
\description{
This function calculates quantile covariance between two variables
using the notion of quantile covariance developed by Li et al. in 2015
}
\examples{
X_0=MASS::mvrnorm(100, mu=rep(0,100), Sigma=diag(100))
 beta=rep(1,100)
 Y_0= X_0\%*\%beta+rnorm(100)
 qcov= qcov_li(X_0, Y_0, 0.5)
}
