% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{signflip_svd}
\alias{signflip_svd}
\title{Sign flip (based on fPQR code from Civieta et al.)}
\usage{
signflip_svd(u, v)
}
\arguments{
\item{u}{first column of U from singular value decomposition UDV^t}

\item{v}{first row of V from singular value decomposition UDV^t}
}
\value{
List of two items
\itemize{
  \item  potentially sign flipped u.
  \item  potentially sign flipped v .
 }
}
\description{
This function flips the signs of u and v to ensure consistency among different solutions for singular value decomposition, and is based off of code written by Civieta et al. in 2021.
}
\examples{
X=MASS::mvrnorm(100, mu=rep(0,100), Sigma=diag(100))
svd_X=svd(X)
col_weight=svd_X$u[,1]
row_weight=svd_X$v[1,]
flip_weights=signflip_svd(col_weight, row_weight)
}
